backend/requirements.txt
asgiref==3.7.2
Django==4.1.10
django-cors-headers==4.2.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.1
mongoengine==0.27.0
pymongo==3.12.3
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.2.4
typing_extensions==4.7.1
tzdata==2023.3
